Refactor flask code from Google code
