+++ tests/test_auth_basic.py
@@ -39,7 +39,7 @@ class TestAll(unittest.TestCase):
         lsd = auth.logout_service_description()
         self.assertEqual(lsd['profile'], 'http://iiif.io/api/auth/0/logout')
         self.assertEqual(lsd['@id'], 'xyz')
-        self.assertEqual(lsd['label'], 'Logout from image server')
+        self.assertEqual(lsd['label'], 'Logout from image server (basic auth)')
 
     def test03_info_authn(self):
         """Test info_authn."""
@@ -53,7 +53,7 @@ class TestAll(unittest.TestCase):
         with dummy_app.test_request_context('/a_request'):
             auth = IIIFAuthBasic()
             ia = auth.image_authn()
-            self.assertEqual(ia, '')
+            self.assertEqual(ia, False)
 
     def test05_login_handler(self):
         """Test login_handler."""
@@ -79,7 +79,7 @@ class TestAll(unittest.TestCase):
             set_cookie = response.headers['Set-Cookie']
             self.assertTrue(
                 re.search(
-                    auth.auth_cookie_name +
+                    auth.account_cookie_name +
                     '=valid-http-basic-login',
                     set_cookie))
         # add bad login params and check fail
@@ -103,49 +103,3 @@ class TestAll(unittest.TestCase):
                     r'<script>window.close\(\);</script>',
                     html))
 
-    def test07_access_token_handler(self):
-        """Test access_token_handler."""
-        with dummy_app.test_request_context('/a_request'):
-            auth = IIIFAuthBasic()
-            response = auth.access_token_handler()
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(
-                response.headers['Content-type'],
-                'application/json')
-            # get_data is bytes in python3
-            j = json.loads(response.get_data().decode('utf-8'))
-            self.assertEqual(
-                j['description'],
-                "No login details received")
-            self.assertEqual(j['error'], "client_unauthorized")
-        # add Authorization header, check we get token
-        h = Headers()
-        h.add('Authorization', b'Basic ' +
-              base64.b64encode(b'userpass:userpass'))
-        with dummy_app.test_request_context('/a_request', headers=h):
-            auth = IIIFAuthBasic()
-            response = auth.access_token_handler()
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(
-                response.headers['Content-type'],
-                'application/json')
-            j = json.loads(response.get_data().decode('utf-8'))
-            self.assertEqual(j['accessToken'], "secret_token_here")  # FIXME
-            self.assertEqual(j['tokenType'], "Bearer")
-            self.assertEqual(j['expiresIn'], 3600)
-        # add callback but no Authorization header
-        with dummy_app.test_request_context('/a_request?callback=CB'):
-            auth = IIIFAuthBasic()
-            response = auth.access_token_handler()
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(
-                response.headers['Content-type'],
-                'application/javascript')
-            # strip JavaScript wrapper and then check JSON
-            js = response.get_data().decode('utf-8')
-            self.assertTrue(re.match('CB\(.*\);', js))
-            j = json.loads(js.lstrip('CB(').rstrip(');'))
-            self.assertEqual(
-                j['description'],
-                "No login details received")
-            self.assertEqual(j['error'], "client_unauthorized")
